Rewrite Action class Gazebo test passed
